Modularize ShimmerElementsGroup
