+++ packages/experiments/src/components/Shimmer/index.ts
@@ -9,4 +9,4 @@ export * from './ShimmerCircle/ShimmerCircle';
 export * from './ShimmerCircle/ShimmerCircle.types';
 export * from './ShimmerGap/ShimmerGap';
 export * from './ShimmerGap/ShimmerGap.types';
-export * from './ShimmerElementsGroup';
+export * from './ShimmerElementsGroup/ShimmerElementsGroup';
